+++ xchange-core/src/test/java/org/knowm/xchange/utils/PriceUtilsTest.java
@@ -9,7 +9,7 @@ public class PriceUtilsTest {
 
     @Test
     public void tickSizeConvertionFormulaTest(){
-        BigDecimal priceAfterRounding = PriceUtils.tickRounding(new BigDecimal(3502.31111111),0.25);
+        BigDecimal priceAfterRounding = PriceUtils.tickRounding(new BigDecimal(3502.31111111),BigDecimal.valueOf(0.25));
         System.out.println(priceAfterRounding);
         Assert.assertEquals(priceAfterRounding,BigDecimal.valueOf(3502.25));
     }
